+++ index.html
@@ -25,6 +25,6 @@
             <p class="clear">Clear all completed</p>
         </div>
     </div>
-    <script src="script.js"></script>
+    <script src="app.js"></script>
 </body>
 </html>
